Added easier way to input time for `timer` function.
Using the HH:MM:SS method for input is good, but a hassle to type.  Plus
it's rare to have the timer go for more than a few minutes, so there's a
lot of extra typing for a short timer.  Added ability to enter time
using h, m, or s identifiers, and we can just use any combination of
those (e.g. 1h5m30, 5m, 30s, 1h30s, etc.).  This will make it a lot
easier to enter time!